refactor: #22 전역상태 모달 추가
